% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_initialization.R
\name{alpha.initialization}
\alias{alpha.initialization}
\title{Alpha Initialization Function}
\usage{
alpha.initialization(pf.vec, S, keep.alpha, alpha0.comp = "equiv")
}
\arguments{
\item{pf.vec}{Profile vector.}

\item{S}{Number of sources.}

\item{keep.alpha}{Logical indicating whether to keep alpha values the same for all profiles.}

\item{alpha0.comp}{Character specifying the method to compute alpha0 ("equiv" or "ones").}
}
\value{
A list of initialized alpha values for each profile.
}
\description{
Initializes alpha values based on the profile vector and user-defined parameters.
}
\details{
Alpha Initialization Function

\code{alpha.initialization} is a function that initializes alpha values based on
the profile vector and user-defined parameters.


This function initializes alpha values based on the profile vector and user-defined
parameters. It can set alpha values uniformly across profiles or customize them
based on the profile's characteristics.
}
\seealso{
\code{\link{getBlockSamples}} for details on extracting block samples.
}
