% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_suitable_L.R
\name{beta.suitable.L}
\alias{beta.suitable.L}
\title{Suitable Beta Computation Function}
\usage{
beta.suitable.L(
  beta,
  lambda,
  function.L,
  gradient.L,
  L.min,
  omega,
  optimization,
  L.step,
  maxIter,
  tol,
  p = NULL,
  gamma = 1,
  range_beta = NULL
)
}
\arguments{
\item{beta}{Initial beta vector.}

\item{lambda}{Regularization parameter.}

\item{function.L}{A function to calculate the objective value.}

\item{gradient.L}{A function to calculate the gradient vector.}

\item{L.min}{Minimum regularization parameter (L).}

\item{omega}{Omega parameter.}

\item{optimization}{Optimization method ("reg" or "cons").}

\item{L.step}{Learning step size for adjusting L.}

\item{maxIter}{Maximum number of iterations.}

\item{tol}{Tolerance level for convergence.}

\item{p}{Number of sources (optional).}

\item{gamma}{Tuning parameter (optional).}
}
\value{
A suitable beta vector for the given lambda value and optimization method.
}
\description{
Computes a suitable beta vector for a given lambda value and optimization method by
iteratively adjusting the regularization parameter.
}
\details{
Suitable Beta Computation Function

\code{beta.suitable.L} is a function that computes a suitable beta vector for a given lambda
value and optimization method.


This function iteratively computes a suitable beta vector for a given lambda value
and optimization method. It adjusts the regularization parameter L and computes the
beta vector that minimizes the objective function within the given tolerance.
}
\seealso{
\code{\link{prox.grad.method}}, \code{\link{norm.proj.method}} for details on the optimization methods.
}
