% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm_predict.R
\name{bwm.predict}
\alias{bwm.predict}
\title{Predict Function for Block-Wise Missing Data Model}
\usage{
bwm.predict(model, X, p, verbose = FALSE)
}
\arguments{
\item{model}{A trained block-wise missing data model.}

\item{X}{Data matrix for prediction.}

\item{p}{Number of sources.}
}
\value{
A vector of predicted outcomes.
}
\description{
Makes predictions using a block-wise missing data model based on the provided model and data.
}
\details{
Predict Function for Block-Wise Missing Data Model

\code{bwm.predict} is a function that makes predictions using a block-wise missing data model.


This function takes a trained block-wise missing data model, input data matrix \code{X}, and the
number of sources \code{p} to make predictions. It considers the profile vector of \code{X}
and computes predictions based on the model's parameters, including alpha, beta, and profiles.
}
\seealso{
\code{\link{get_profile}}, \code{\link{as.binary}} for details on profile extraction.
}
