% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm_train.R
\name{bwm.train}
\alias{bwm.train}
\title{Block-Wise Missing Data Model Training Function}
\usage{
bwm.train(
  p,
  X,
  y,
  lambda,
  L.step = 1.5,
  maxIter = 300,
  tol = 1e-12,
  omega.alpha = "LR",
  tol.alpha = 1e-12,
  maxIter.alpha = 20,
  omega.beta = "LR",
  alpha0.comp = "equiv",
  tol.beta = 1e-12,
  maxIter.beta = 20,
  gamma = 1,
  to.normalize = F,
  beta0,
  alpha0,
  verbose = TRUE,
  maxClasses = 20
)
}
\arguments{
\item{p}{Number of sources.}

\item{X}{Data matrix for training.}

\item{y}{Response vector for training.}

\item{lambda}{Vector of regularization parameters.}

\item{L.step}{Learning step size for L.}

\item{maxIter}{Maximum number of iterations for training.}

\item{tol}{Tolerance level for convergence.}

\item{omega.alpha}{Alpha regularization method ("LR" or other).}

\item{tol.alpha}{Tolerance level for alpha convergence.}

\item{maxIter.alpha}{Maximum number of iterations for alpha optimization.}

\item{omega.beta}{Beta regularization method ("LR" or other).}

\item{alpha0.comp}{Method for alpha initialization ("equiv" or other).}

\item{tol.beta}{Tolerance level for beta convergence.}

\item{maxIter.beta}{Maximum number of iterations for beta optimization.}

\item{gamma}{Tuning parameter for optimization.}

\item{to.normalize}{Logical indicating whether to normalize the features.}

\item{beta0}{Optional initial beta values (list or vector).}

\item{alpha0}{Optional initial alpha values (list or vector).}

\item{maxClasses}{Optional number of max classes to classify (int).}
}
\value{
A list containing the trained block-wise missing data model parameters.
}
\description{
Trains a block-wise missing data model using the provided data and user-defined parameters.
}
\details{
Block-Wise Missing Data Model Training Function

\code{bwm.train} is a function that trains a block-wise missing data model using the provided data


This function trains a block-wise missing data model using the provided data and user-defined
parameters. It supports various options for regularization, initialization, and optimization.
}
\seealso{
\code{\link{alpha.initialization}}, \code{\link{beta.initialization}},
\code{\link{prox.grad.iter.method}} for details on initialization and optimization methods.
}
