% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const_Lipschitz_alpha_ce.R
\name{compute_lipschitz_alpha.ce}
\alias{compute_lipschitz_alpha.ce}
\title{Compute Lipschitz Constant for Alpha Optimization}
\usage{
compute_lipschitz_alpha.ce(X.m, n.m)
}
\arguments{
\item{X.m}{A numeric matrix containing the feature values for the profile.}

\item{n.m}{The number of samples corresponding to the current profile.}
}
\value{
The Lipschitz constant for alpha optimization, which is the maximum of the calculated values
across all sources.
}
\description{
This function computes the Lipschitz constant that is used in the alpha optimization process.
The Lipschitz constant is calculated based on the singular values of the feature matrix (X.m)
and is important for controlling the convergence behavior during optimization.
}
\details{
Compute Lipschitz Constant for Alpha Optimization

\code{compute_lipschitz_alpha.ce} calculates the Lipschitz constant for alpha optimization
based on the provided tilde beta values and feature matrix.


The function computes the Lipschitz constant by calculating the singular values of the input
feature matrix (\code{X.m}) using Singular Value Decomposition (SVD). The largest singular value
is squared and then divided by a factor of 4 times the number of samples (\code{n.m}). The function
returns the computed value.
}
\examples{
# Example usage:
X.m <- matrix(rnorm(50), nrow = 10, ncol = 5)  # Example feature matrix
n.m <- 10  # Number of samples
lipschitz_constant <- compute_lipschitz_alpha.ce(X.m, n.m)
print(lipschitz_constant)

}
