% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const_Lipschitz_alpha.R
\name{const.Lipschitz.alpha}
\alias{const.Lipschitz.alpha}
\title{Compute Lipschitz Constant for Alpha Optimization}
\usage{
const.Lipschitz.alpha(tilde.beta)
}
\arguments{
\item{tilde.beta}{Matrix of tilde beta values.}
}
\value{
The Lipschitz constant for alpha optimization.
}
\description{
Computes the Lipschitz constant for alpha optimization using the provided tilde beta values.
}
\details{
Compute Lipschitz Constant for Alpha Optimization

\code{const.Lipschitz.alpha} is a function that computes the Lipschitz constant for
alpha optimization based on the provided tilde beta values.


This function calculates the Lipschitz constant used in the alpha optimization process.
It takes a matrix of tilde beta values as input and computes the maximum sum of squares
of columns in the matrix as the Lipschitz constant.
}
