% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_log.R
\name{f.log}
\alias{f.log}
\title{Logistic Loss Function Evaluation Function}
\usage{
f.log(y.m, alpha.m, tilde.beta)
}
\arguments{
\item{y.m}{Observed values.}

\item{alpha.m}{Alpha values.}

\item{tilde.beta}{Tilde beta values.}
}
\value{
The value of the logistic loss function evaluated with the given parameters.
}
\description{
Evaluates the logistic loss function for a given set of parameters, including observed values (y.m),
alpha values (alpha.m), and tilde beta values (tilde.beta).
}
\details{
Logistic Loss Function Evaluation Function

\code{f.log} is a function that evaluates the logistic loss function for a given set of parameters.


This function computes the value of the logistic loss function for a given set of parameters.
}
