% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ls.R
\name{f.ls}
\alias{f.ls}
\title{Least-Square Loss Function Evaluation Function}
\usage{
f.ls(y.m, alpha.m, tilde.beta)
}
\arguments{
\item{y.m}{Observed values.}

\item{alpha.m}{Alpha values.}

\item{tilde.beta}{Tilde beta values.}
}
\value{
The value of the least-square loss function evaluated with the given parameters.
}
\description{
Evaluates the least-square loss function for a given set of parameters, including observed values (y.m),
alpha values (alpha.m), and tilde beta values (tilde.beta).
}
\details{
Least-Square Loss Function Evaluation Function

\code{f.ls} is a function that evaluates the least-square loss function for a given set of parameters.


This function computes the value of the least-square loss function for a given set of parameters.
}
