% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Loss Function Evaluation Function}
\usage{
f(y.m, alpha.m, tilde.beta, loss.function)
}
\arguments{
\item{y.m}{Observed values.}

\item{alpha.m}{Alpha values.}

\item{tilde.beta}{Tilde beta values.}

\item{loss.function}{Type of loss function ("ls" for least-square, "log" for logistic).}
}
\value{
The value of the loss function evaluated with the given parameters.
}
\description{
Evaluates a loss function for a given set of parameters, including observed values (y.m),
alpha values (alpha.m), tilde beta values (tilde.beta), and the loss function type.
}
\details{
Loss Function Evaluation Function

\code{f} is a function that evaluates a loss function for a given set of parameters.


This function computes the value of a loss function for a given set of parameters based on
the specified loss function type. It supports both least-square (ls) and logistic (log) loss
functions.
}
\seealso{
\code{\link{f.ls}}, \code{\link{f.log}} for details on specific loss functions.
}
