% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Objective Function Evaluation Function}
\usage{
g(p, X, y, alpha, beta, pf.vec, loss.function)
}
\arguments{
\item{p}{A numeric vector representing the number of sources. This can be used for specifying the number of blocks or sources in the model.}

\item{X}{A numeric matrix containing the data, with rows representing samples and columns representing features.}

\item{y}{A numeric vector representing the response or target values corresponding to each sample.}

\item{alpha}{A numeric vector or list representing the alpha values for the model, used in optimization.}

\item{beta}{A numeric vector representing the beta values or coefficients in the model.}

\item{pf.vec}{A vector indicating the profile or groupings for each sample, used to organize the data into different profiles.}

\item{loss.function}{A character string specifying the type of loss function to be used: "ls" for least-squares or "log" for logistic regression.}
}
\value{
The value of the objective function, evaluated using the provided parameters and the specified loss function type.
This can be a scalar value representing the performance of the model with respect to the data and the current parameters.
}
\description{
This function serves as a wrapper for evaluating the objective function with a given set of parameters.
It takes in the number of sources, the data matrix, the response vector, the alpha and beta values,
and the profile vector, then computes the objective function using the specified loss function.
}
\details{
Objective Function Evaluation Wrapper

\code{g} is a wrapper function that evaluates the objective function for a given set of parameters.


The function evaluates the objective function by calling the \code{objective.fun} function. It requires the following parameters:
the number of sources (\code{p}), the data matrix (\code{X}), the response vector (\code{y}), the alpha vector (\code{alpha}),
the beta vector (\code{beta}), the profile vector (\code{pf.vec}), and the loss function type (\code{loss.function}).
Depending on the loss function ("ls" or "log"), it will compute the appropriate objective function for regression or classification tasks.
}
\examples{
# Example usage:
p <- c(2, 3)  # Number of sources
X <- matrix(rnorm(50), nrow = 10, ncol = 5)  # Data matrix
y <- c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0)  # Response vector
alpha <- c(0.1, 0.2, 0.3)  # Alpha values
beta <- c(0.5, -0.2, 0.3, 0.1, -0.4)  # Beta values
pf.vec <- c(1, 1, 2, 2, 3, 3, 4, 4, 5, 5)  # Profile vector
loss.function <- "log"  # Loss function type
obj_value <- g(p, X, y, alpha, beta, pf.vec, loss.function)
print(obj_value)

}
\seealso{
\code{\link{objective.fun}} for the details of how the objective function is computed based on the parameters.
}
