% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile.R
\name{get_profile}
\alias{get_profile}
\title{Generate Profile Vector from Data Matrix}
\usage{
get_profile(p, X, showTable = F, saveTable = F)
}
\arguments{
\item{p}{Source vector representing the number of variables in each source.}

\item{X}{Data matrix.}

\item{showTable}{Logical, whether to print the profile matrix (default is FALSE).}

\item{saveTable}{Logical, whether to save the profile matrix as a CSV file (default is FALSE).}
}
\value{
A profile vector representing profiles for each sample in the data matrix.
}
\description{
Generates a profile vector from a data matrix based on the provided source vector. Each profile is
represented as a binary number in the profile vector.
}
\details{
Generate Profile Vector from Data Matrix

\code{get_profile} is a function that generates a profile vector from a data matrix based on the provided
source vector.


This function generates a profile vector based on the provided source vector and data matrix. The profile
vector represents profiles for each sample in the data matrix using binary numbers. Each binary digit
corresponds to whether a variable is present in a profile.
}
\seealso{
\code{\link{get_profile}} for details on how profiles are generated.
}
