% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlockSamples.R
\name{getBlockSamples}
\alias{getBlockSamples}
\title{Get Block Samples for a Given Profile}
\usage{
getBlockSamples(pf.vec, m, S)
}
\arguments{
\item{pf.vec}{Profile vector containing profiles for each sample.}

\item{m}{Profile of interest.}

\item{S}{Number of sources.}
}
\value{
A list containing the samples and sources associated with the specified profile.
}
\description{
Retrieves the samples and sources associated with a given profile from the profile vector.
}
\details{
Get Block Samples for a Given Profile

\code{getBlockSamples} is a function that retrieves the samples and sources associated with a given profile.


This function identifies and collects the samples and sources associated with a specific profile
from the profile vector. It returns the samples belonging to the profile and the sources that are
relevant to that profile.
}
