% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_f_log.R
\name{gradient.f.log}
\alias{gradient.f.log}
\title{Gradient of the Logistic Objective Function Evaluation Function}
\usage{
gradient.f.log(y.m, alpha.m, tilde.beta)
}
\arguments{
\item{y.m}{Observed values.}

\item{alpha.m}{Alpha values.}

\item{tilde.beta}{Tilde beta values.}
}
\value{
The gradient of the logistic objective function evaluated with the given parameters.
}
\description{
Computes the gradient of the logistic objective function for a given set of parameters, including observed values (y.m),
alpha values (alpha.m), and tilde beta values (tilde.beta).
}
\details{
Gradient of the Logistic Objective Function Evaluation Function

\code{gradient.f.log} is a function that computes the gradient of the logistic objective function for a given set of parameters.


This function computes the gradient of the logistic objective function for a given set of parameters. The gradient is
computed with respect to the alpha values.
}
