% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_f_ls.R
\name{gradient.f.ls}
\alias{gradient.f.ls}
\title{Gradient of the Least-Square Objective Function Evaluation Function}
\usage{
gradient.f.ls(y.m, alpha.m, tilde.beta)
}
\arguments{
\item{y.m}{Observed values.}

\item{alpha.m}{Alpha values.}

\item{tilde.beta}{Tilde beta values.}
}
\value{
The gradient of the least-square objective function evaluated with the given parameters.
}
\description{
Computes the gradient of the least-square objective function for a given set of parameters, including observed values (y.m),
alpha values (alpha.m), and tilde beta values (tilde.beta).
}
\details{
Gradient of the Least-Square Objective Function Evaluation Function

\code{gradient.f.ls} is a function that computes the gradient of the least-square objective function for a given set of parameters.


This function computes the gradient of the least-square objective function for a given set of parameters. The gradient is
computed with respect to the alpha values.
}
