% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_f.R
\name{gradient.f}
\alias{gradient.f}
\title{Gradient of the Objective Function Evaluation Function}
\usage{
gradient.f(y.m, alpha.m, tilde.beta, loss.function)
}
\arguments{
\item{y.m}{Observed values.}

\item{alpha.m}{Alpha values.}

\item{tilde.beta}{Tilde beta values.}

\item{loss.function}{Type of loss function ("ls" for least-square, "log" for logistic).}
}
\value{
The gradient of the objective function evaluated with the given parameters.
}
\description{
Computes the gradient of the objective function for a given set of parameters, including observed values (y.m),
alpha values (alpha.m), tilde beta values (tilde.beta), and loss function type.
}
\details{
Gradient of the Objective Function Evaluation Function

\code{gradient.f} is a function that computes the gradient of the objective function for a given set of parameters.


This function computes the gradient of the objective function for a given set of parameters and loss function type.
}
