% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_g_ls.R
\name{gradient.g.ls}
\alias{gradient.g.ls}
\title{Gradient of the Least-Square Objective Function Evaluation Function (Generalized)}
\usage{
gradient.g.ls(p, X, y, alpha, beta, pf.vec)
}
\arguments{
\item{p}{Source vector specifying the number of variables per source.}

\item{X}{Data matrix.}

\item{y}{Observed values.}

\item{alpha}{Alpha values.}

\item{beta}{Beta values.}

\item{pf.vec}{Profile vector containing profiles for each sample.}
}
\value{
The gradient of the least-square objective function evaluated with the given parameters.
}
\description{
Computes the gradient of the least-square objective function for a given set of parameters, including source vector (p), data matrix (X),
observed values (y), alpha values (alpha), beta values (beta), and profile vector (pf.vec).
}
\details{
Gradient of the Least-Square Objective Function Evaluation Function (Generalized)

\code{gradient.g.ls} is a function that computes the gradient of the least-square objective function for a given set of parameters in a generalized manner, considering multiple profiles.


This function computes the gradient of the least-square objective function for a given set of parameters. The gradient is
computed with respect to the beta values, considering multiple profiles and alpha values. It generalizes the gradient computation for least-square loss.
}
