% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_norm_log.R
\name{inner.norm.log}
\alias{inner.norm.log}
\title{Inner Normalization Function (Generalized)}
\usage{
inner.norm.log(y.m, alpha.m, tilde.beta, l)
}
\arguments{
\item{y.m}{Observed values.}

\item{alpha.m}{Alpha values.}

\item{tilde.beta}{Beta values.}

\item{l}{Index for the current observation.}
}
\value{
The inner normalization term evaluated with the given parameters.
}
\description{
Computes the inner normalization term for a given set of parameters, including observed values (y.m), alpha values (alpha.m), and beta values (tilde.beta).
}
\details{
Inner Normalization Function (Generalized)

\code{inner.norm.log} is a function that computes the inner normalization term for a given set of parameters in a generalized manner, considering multiple sources.


This function computes the inner normalization term for a given set of parameters, considering multiple sources and alpha values.
}
