% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_min.R
\name{L.min}
\alias{L.min}
\title{Minimum Lipschitz Value Calculation Function}
\usage{
L.min(beta.current, beta.prev, gradient)
}
\arguments{
\item{beta.current}{The current parameter vector.}

\item{beta.prev}{The previous parameter vector.}

\item{gradient}{A function that computes the gradient of a specific objective function.}
}
\value{
The minimum Lipschitz constant evaluated with the given parameters and gradients.
}
\description{
Calculates the minimum Lipschitz constant for a given set of parameters (\code{beta.current} and \code{beta.prev}) and their gradients (\code{gradient}).
}
\details{
Minimum Lipschitz Value Calculation Function

\code{L.min} is a function that calculates the minimum Lipschitz constant for a given set of parameters and gradients.


This function calculates the minimum Lipschitz constant, which is used in optimization algorithms to determine step sizes.
}
