% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_computation_l2.R
\name{mu_computation.l2}
\alias{mu_computation.l2}
\title{Compute Regularization Parameter mu for L2 Regularization}
\usage{
mu_computation.l2(beta, lambda)
}
\arguments{
\item{beta}{The current parameter vector.}

\item{lambda}{The regularization strength.}
}
\value{
The computed regularization parameter \eqn{\mu} for L2 regularization.
}
\description{
Computes the regularization parameter \eqn{\mu} for L2 regularization based on the current parameter vector (\code{beta}) and the regularization strength (\code{lambda}).
}
\details{
Compute Regularization Parameter mu for L2 Regularization

\code{mu_computation.l2} is a function that computes the regularization parameter \eqn{\mu} for L2 regularization.


This function calculates the regularization parameter \eqn{\mu} for L2 regularization based on the input parameter vector \code{beta} and the regularization strength \code{lambda}.
}
\examples{
# Example usage:
beta <- c(0.5, -0.7, 0.2, 0.1)
lambda <- 0.4
mu <- mu_computation.l2(beta, lambda)

}
