% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_proj_method.R
\name{norm.proj.method}
\alias{norm.proj.method}
\title{Projection onto Regularized Norm Ball}
\usage{
norm.proj.method(beta, lambda, L, gradient, omega, tol = 0.001)
}
\arguments{
\item{beta}{The parameter vector to be projected.}

\item{lambda}{The regularization strength.}

\item{L}{The Lipschitz constant.}

\item{gradient}{The gradient of the objective function.}

\item{omega}{The type of norm ball to project onto ("LR" for L1 norm, "RR" for L2^2 norm).}

\item{tol}{Tolerance for numerical comparisons (default is 1e-3).}
}
\value{
The projected parameter vector.
}
\description{
This function performs the projection of a parameter vector \code{beta} onto a regularized norm ball, where the type of norm ball (L1 or L2) and the regularization strength (\code{lambda}) are determined by the user.
}
\details{
Projection onto Regularized Norm Ball

\code{norm.proj.method} is a function that performs the projection of a parameter vector onto a regularized norm ball.


This function calculates the projection of a parameter vector \code{beta} onto a regularized norm ball. The type of norm ball (L1 or L2) and the regularization strength (\code{lambda}) are specified by the user. It uses the Lipschitz constant (\code{L}), gradient, and tolerance (\code{tol}) for the projection.
}
\examples{
# Example usage:
beta <- c(0.5, -0.7, 0.2, 0.1)
lambda <- 0.4
L <- 0.1
gradient <- c(0.1, 0.2, -0.3, 0.4)
omega <- "LR"
tol <- 1e-3
projected_beta <- norm.proj.method(beta, lambda, L, gradient, omega, tol)

}
