% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_fun_log.R
\name{objective.fun.log}
\alias{objective.fun.log}
\title{Objective Function Evaluation for Logistic Loss}
\usage{
objective.fun.log(p, X, y, beta, alpha, pf.vec)
}
\arguments{
\item{p}{A vector specifying the number of variables for each source.}

\item{X}{The feature matrix.}

\item{y}{The target vector.}

\item{beta}{The parameter vector.}

\item{alpha}{A list of alpha weights.}

\item{pf.vec}{The profile vector.}
}
\value{
The value of the logistic loss objective function.
}
\description{
This function evaluates the objective function for a logistic loss model based on the given model parameters, data, and profile vector.
}
\details{
Objective Function Evaluation for Logistic Loss

\code{objective.fun.log} is a function that evaluates the objective function for a logistic loss model.


This function calculates the value of the objective function for a logistic loss model. The objective function is used to measure the goodness-of-fit of the model to the data.
}
\examples{
# Example usage:
p <- c(2, 3)
X <- matrix(rnorm(100), ncol = sum(p))
y <- c(0, 1, 0, 1, 0)
beta <- c(0.5, -0.7, 0.2, 0.1, -0.3)
alpha <- list(c(0.1, 0.2), c(0.3, 0.4, 0.5))
pf.vec <- as.factor(c(1, 2, 1, 3, 2))
objective <- objective.fun.log(p, X, y, beta, alpha, pf.vec)

}
