% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_fun.R
\name{objective.fun}
\alias{objective.fun}
\title{Objective Function Evaluation}
\usage{
objective.fun(p, X, y, beta, alpha, pf.vec, loss.function)
}
\arguments{
\item{p}{A vector specifying the number of variables for each source.}

\item{X}{The feature matrix.}

\item{y}{The target vector.}

\item{beta}{The parameter vector.}

\item{alpha}{A list of alpha weights.}

\item{pf.vec}{The profile vector.}

\item{loss.function}{The loss function ("ls" for least-square or "log" for logistic or "ce" for cross-entropy).}
}
\value{
The value of the objective function.
}
\description{
This function evaluates the objective function for a given model, loss function, and data.
}
\details{
Objective Function Evaluation

\code{objective.fun} is a function that evaluates the objective function for a given model and loss function.


This function calculates the value of the objective function based on the given model parameters, data, and loss function. The objective function is used to measure the goodness-of-fit of the model to the data.
}
\examples{
# Example usage:
p <- c(2, 3)
X <- matrix(rnorm(100), ncol = sum(p))
y <- rnorm(100)
beta <- c(0.5, -0.7, 0.2, 0.1, -0.3)
alpha <- list(c(0.1, 0.2), c(0.3, 0.4, 0.5))
pf.vec <- as.factor(c(1, 2, 1, 3, 2))
loss.function <- "ls"
objective <- objective.fun(p, X, y, beta, alpha, pf.vec, loss.function)

}
