% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_norm_proj_method.R
\name{omega.norm.proj.method}
\alias{omega.norm.proj.method}
\title{Omega-Norm Projection Method}
\usage{
omega.norm.proj.method(
  y.m,
  alpha0,
  tilde.beta,
  omega,
  L.step,
  maxIter,
  tol,
  loss.function
)
}
\arguments{
\item{y.m}{The target vector for a specific profile.}

\item{alpha0}{The initial alpha vector.}

\item{tilde.beta}{The parameter matrix.}

\item{omega}{The omega value (either "LR" or "RR").}

\item{L.step}{The L-step factor.}

\item{maxIter}{The maximum number of iterations.}

\item{tol}{The tolerance level for convergence.}

\item{loss.function}{The loss function ("ls" or "log").}
}
\value{
The computed alpha vector.
}
\description{
This function applies the omega-norm projection method to compute the alpha vector based on the given parameters and data.
}
\details{
Omega-Norm Projection Method

\code{omega.norm.proj.method} is a function that performs the omega-norm projection method to compute the alpha vector.


The omega-norm projection method is used to compute the alpha vector in a model. This function iteratively updates the alpha vector until convergence or reaching the maximum number of iterations.
}
\examples{
# Example usage:
y.m <- c(0, 1, 0, 1, 0)
alpha0 <- c(0.1, 0.2)
tilde.beta <- matrix(c(0.5, -0.7, 0.2, 0.1, -0.3), ncol = 2)
omega <- "LR"
L.step <- 1.5
maxIter <- 100
tol <- 1e-6
loss.function <- "log"
computed_alpha <- omega.norm.proj.method(y.m, alpha0, tilde.beta, omega, L.step, maxIter, tol, loss.function)

}
