% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_grad_iter_method.R
\name{prox.grad.iter.method}
\alias{prox.grad.iter.method}
\title{Proximal Gradient Iteration Method}
\usage{
prox.grad.iter.method(
  p,
  X,
  y,
  alpha,
  beta0,
  pf.vec,
  lambda,
  omega,
  L.step,
  maxIter,
  tol,
  gamma,
  loss.function
)
}
\arguments{
\item{p}{The source vector.}

\item{X}{The data matrix.}

\item{y}{The target vector.}

\item{alpha}{The alpha vector.}

\item{beta0}{The initial beta vector.}

\item{pf.vec}{The profile vector.}

\item{lambda}{The lambda regularization parameter.}

\item{omega}{The omega value (either "LR" or "RR").}

\item{L.step}{The L-step factor.}

\item{maxIter}{The maximum number of iterations.}

\item{tol}{The tolerance level for convergence.}

\item{gamma}{The gamma value.}

\item{loss.function}{The loss function ("ls" or "log").}
}
\value{
The computed beta vector.
}
\description{
This function performs proximal gradient iterations to compute the beta vector based on the given parameters and data.
}
\details{
Proximal Gradient Iteration Method

\code{prox.grad.iter.method} is a function that iteratively computes the beta vector using the proximal gradient method.


The proximal gradient method is used to compute the beta vector iteratively. This function updates the beta vector until convergence or reaching the maximum number of iterations.
}
\examples{
# Example usage:
p <- c(2, 3)
X <- matrix(runif(100), ncol = 10)
y <- rnorm(10)
alpha <- c(0.1, 0.2)
beta0 <- rep(0, 10)
pf.vec <- factor(rep(1:5, each = 2))
lambda <- 0.5
omega <- "LR"
L.step <- 1.5
maxIter <- 100
tol <- 1e-6
gamma <- 0.01
loss.function <- "ls"
computed_beta <- prox.grad.iter.method(p, X, y, alpha, beta0, pf.vec, lambda, omega, L.step, maxIter, tol, gamma, loss.function)

}
