% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_grad_method.R
\name{prox.grad.method}
\alias{prox.grad.method}
\title{Proximal Gradient Method}
\usage{
prox.grad.method(beta, lambda, L, gradient, omega, p, gamma)
}
\arguments{
\item{beta}{The input vector.}

\item{lambda}{The regularization parameter.}

\item{L}{The Lipschitz constant.}

\item{gradient}{The gradient function.}

\item{omega}{The regularization type ("LR", "RR", "EN", or "GL").}

\item{p}{The source vector (optional, required for "GL").}

\item{gamma}{The gamma value (optional, required for "EN" and "GL").}
}
\value{
The computed proximal operator.
}
\description{
This function computes the proximal operator of a regularization term based on the given parameters.
}
\details{
Proximal Gradient Method

\code{prox.grad.method} is a function that computes the proximal operator of a regularization term.


The proximal gradient method is used to compute the proximal operator of a regularization term. The function returns the proximal operator based on the specified regularization type.
}
\examples{
# Example usage:
beta <- c(0.5, -0.2, 0.7)
lambda <- 0.1
L <- 1.0
gradient <- function(x) 2 * x  # Example gradient function
omega <- "LR"
computed_prox <- prox.grad.method(beta, lambda, L, gradient, omega)

}
