% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operator_l1.R
\name{prox.operator.l1}
\alias{prox.operator.l1}
\title{Proximal Operator for L1 Regularization}
\usage{
prox.operator.l1(u, mu)
}
\arguments{
\item{u}{The input vector.}

\item{mu}{The regularization parameter.}
}
\value{
The computed proximal operator for L1 regularization.
}
\description{
This function computes the proximal operator for L1 regularization based on the given parameters.
}
\details{
Proximal Operator for L1 Regularization

\code{prox.operator.l1} is a function that computes the proximal operator for L1 regularization.


The proximal operator for L1 regularization (also known as soft-thresholding) is used to perform shrinkage on the elements of the input vector u. It applies the element-wise soft-thresholding operation to u with a threshold of mu.
}
\examples{
# Example usage:
u <- c(0.5, -0.2, 0.7)
mu <- 0.1
prox <- prox.operator.l1(u, mu)

}
