% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operator_l1_2.R
\name{prox.operator.l1_l2}
\alias{prox.operator.l1_l2}
\title{Proximal Operator for L1/L2 Regularization}
\usage{
prox.operator.l1_l2(p, u, mu)
}
\arguments{
\item{p}{A vector specifying the partition of the input vector u.}

\item{u}{The input vector.}

\item{mu}{The regularization parameter.}
}
\value{
The computed proximal operator for L1/L2 regularization.
}
\description{
This function computes the proximal operator for L1/L2 regularization based on the given parameters.
}
\details{
Proximal Operator for L1/L2 Regularization

\code{prox.operator.l1_l2} is a function that computes the proximal operator for L1/L2 regularization.


The proximal operator for L1/L2 regularization is used to perform regularization on groups of elements of the input vector u, where each group is defined by the partition vector p. It applies the element-wise proximal operator for L1 regularization within each group.
}
\examples{
# Example usage:
p <- c(2, 3, 2)
u <- c(0.5, -0.2, 0.7, 1.0, -1.5, 2.0, -0.3, 0.8)
mu <- 0.1
prox <- prox.operator.l1_l2(p, u, mu)

}
