% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operator_l2.R
\name{prox.operator.l2}
\alias{prox.operator.l2}
\title{Proximal Operator for L2 Regularization}
\usage{
prox.operator.l2(u, mu)
}
\arguments{
\item{u}{The input vector.}

\item{mu}{The regularization parameter for L2 regularization.}
}
\value{
The computed proximal operator for L2 (ridge) regularization.
}
\description{
This function computes the proximal operator for L2 (ridge) regularization based on the given parameters.
}
\details{
Proximal Operator for L2 Regularization

\code{prox.operator.l2} is a function that computes the proximal operator for L2 (ridge) regularization.


The proximal operator for L2 regularization, also known as ridge regularization, applies a shrinkage operation to the input vector u by dividing it by (1 + mu).
}
\examples{
# Example usage:
u <- c(0.5, -0.2, 0.7, 1.0, -1.5, 2.0, -0.3, 0.8)
mu <- 0.1
prox <- prox.operator.l2(u, mu)

}
