% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Load Required R Packages}
\usage{
setup(verbose = TRUE)
}
\description{
This function loads specific R packages that are required for the main algorithm to run successfully.
}
\details{
Load Required R Packages

\code{setup} is a function that loads the necessary R packages for the execution of the main algorithm.


The function loads the following R packages: caret, rlang, MASS, corrplot, glmnet, binaryLogic, ade4, naniar, mvtnorm, devtools, factoextra, and nnet.
}
\examples{
# Example usage:
setup()

}
